Move namespace of native contracts (#2137)
