+++ src/neo/SmartContract/Native/GasToken.cs
@@ -2,7 +2,7 @@ using Neo.Cryptography.ECC;
 using Neo.Ledger;
 using Neo.Network.P2P.Payloads;
 
-namespace Neo.SmartContract.Native.Tokens
+namespace Neo.SmartContract.Native
 {
     public sealed class GasToken : Nep17Token<AccountState>
     {
