+++ src/neo/SmartContract/Native/NativeContract.cs
@@ -1,8 +1,5 @@
 using Neo.IO;
 using Neo.SmartContract.Manifest;
-using Neo.SmartContract.Native.Designate;
-using Neo.SmartContract.Native.Oracle;
-using Neo.SmartContract.Native.Tokens;
 using Neo.VM;
 using Neo.VM.Types;
 using System;
@@ -27,7 +24,7 @@ namespace Neo.SmartContract.Native
         public static GasToken GAS { get; } = new GasToken();
         public static PolicyContract Policy { get; } = new PolicyContract();
         public static OracleContract Oracle { get; } = new OracleContract();
-        public static DesignateContract Designate { get; } = new DesignateContract();
+        public static DesignationContract Designation { get; } = new DesignationContract();
 
         public abstract string Name { get; }
         public byte[] Script { get; }
