+++ tests/neo.UnitTests/Network/P2P/Payloads/UT_Transaction.cs
@@ -7,7 +7,6 @@ using Neo.Ledger;
 using Neo.Network.P2P.Payloads;
 using Neo.SmartContract;
 using Neo.SmartContract.Native;
-using Neo.SmartContract.Native.Tokens;
 using Neo.VM;
 using Neo.Wallets;
 using System;
