+++ tests/neo.UnitTests/SmartContract/Native/UT_DesignationContract.cs
@@ -7,7 +7,6 @@ using Neo.Network.P2P.Payloads;
 using Neo.Persistence;
 using Neo.SmartContract;
 using Neo.SmartContract.Native;
-using Neo.SmartContract.Native.Designate;
 using Neo.UnitTests.Extensions;
 using System.Linq;
 using System.Numerics;
@@ -15,7 +14,7 @@ using System.Numerics;
 namespace Neo.UnitTests.SmartContract.Native
 {
     [TestClass]
-    public class UT_DesignateContract
+    public class UT_DesignationContract
     {
         private StoreView _snapshot;
 
@@ -37,7 +36,7 @@ namespace Neo.UnitTests.SmartContract.Native
             };
             UInt160 committeeMultiSigAddr = NativeContract.NEO.GetCommitteeAddress(snapshot1);
             ECPoint[] validators = NativeContract.NEO.ComputeNextBlockValidators(snapshot1);
-            var ret = NativeContract.Designate.Call(
+            var ret = NativeContract.Designation.Call(
                 snapshot1,
                 new Nep17NativeContractExtensions.ManualWitness(committeeMultiSigAddr),
                 "designateAsRole",
@@ -46,7 +45,7 @@ namespace Neo.UnitTests.SmartContract.Native
             );
             snapshot1.Commit();
             var snapshot2 = _snapshot.Clone();
-            ret = NativeContract.Designate.Call(
+            ret = NativeContract.Designation.Call(
                 snapshot2,
                 "getDesignatedByRole",
                 new ContractParameter(ContractParameterType.Integer) { Value = new BigInteger((int)Role.StateValidator) },
@@ -62,7 +61,7 @@ namespace Neo.UnitTests.SmartContract.Native
             (ret as VM.Types.Array)[5].GetSpan().ToHexString().Should().Be(validators[5].ToArray().ToHexString());
             (ret as VM.Types.Array)[6].GetSpan().ToHexString().Should().Be(validators[6].ToArray().ToHexString());
 
-            ret = NativeContract.Designate.Call(
+            ret = NativeContract.Designation.Call(
                 snapshot2,
                 "getDesignatedByRole",
                 new ContractParameter(ContractParameterType.Integer) { Value = new BigInteger((int)Role.StateValidator) },
